Drop requirement for CLA in CONTRIBUTING.md
It is no longer necessary for contributors to send a CLA to ARM
before making contributions. Contributors must instead add a
"Signed-off-by:" line to each commit, which certifies that the
contribution is made under the Developer Certificate of Origin
(DCO).

Update CONTRIBUTING.md to reflect this new policy and add a copy of
the DCO to the repository.

Fixes projectNe10/Ne10/issues/177

Signed-off-by: David Cunado <david.cunado@arm.com>