fix some CppCheck warnings, mostly related to C89 style code (we use C99)
In C89, variables had to be declared at the start of a code block.
Many programmers though they have to be declared at the beginning
of a function even. by declaring variables where they are first used,
or at least at the beginning of the code block where they are used,
we give the compiler more options for optimization, while also slightly
improving code readability.

Signed-off-by: Robin Vobruba <hoijui.quaero@gmail.com>